Added specs

+++ spock-specs/src/test.java1.8/java/org/spockframework/smoke/mock/ISquare.java
@@ -1,9 +1,9 @@
 package org.spockframework.smoke.mock;
 
 public interface ISquare {
-  double getLength();
+  int getLength();
 
-  default double getArea() {
+  default int getArea() {
     return getLength() * getLength();
   }
 }
